rerefactor: mv UpgradeVector -> ocm.gardener
Considering UpgradeVector (a source and target component-id, where
version is different, and component-name is identital) is not actually
specific for GitHub, but indeed more closely related to OCM (in
particular: operations _on_ OCM), mv into ocm.gardener.
